add server-task, server-common
